web/backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
psycopg2-binary
django-leaflet
Pillow
djangorestframework-simplejwt
drf-spectacular
django-filter
django-simple-history
dj-database-url
whitenoise
gunicorn
python-dotenv
cloudinary
django-cloudinary-storage